# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Template Simbi API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/authors/{id}/books": {
            "get": {
                "tags": ["Authors"],
                "summary": "Lista todos os livros cadastrados na API para um determinado autor",
                "description": "Lista todos os livros cadastrados por um determinado autor.",
                "operationId": "05e289382ad925246b99df38487e708f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do autor",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "2234f840-417e-4944-ac9b-e7e6eb06c590"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Books list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "13b35be6-65a7-4d7f-9ad2-8caaf3c75344"
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Curso avançado de Java"
                                            },
                                            "publisher": {
                                                "type": "string",
                                                "example": "Editora XPTO"
                                            },
                                            "author": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "13b35be6-65a7-4d7f-9ad2-8caaf3c75344"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Autor Y"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "createdAt": {
                                                "type": "string",
                                                "example": "2022-12-14T22:24:26+00:00"
                                            },
                                            "updatedAt": {
                                                "type": "string",
                                                "example": "2022-12-14T22:24:26+00:00"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição com erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bad Request"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Proibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Internal server error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}